imp. implement crud bank for user
