+++ app/(tabs)/(account)/bank-account/bank-account-list/BankAccount.translate.ts
@@ -1,0 +1,27 @@
+const TEXT_TRANSLATE_BANK_ACCOUNT = {
+  TITLE: {
+    ACCOUNT: "Cá nhân",
+    ENTER_ADDRESS: "Nhập địa chỉ",
+    ENTER_FULLNAME: "Nhập họ và tên",
+    ENTER_PHONE_NUMBER: "Nhập số điện thoại",
+    BANK_ACCOUNT: "Tài khoản ngân hàng",
+    ACCOUNT_HOLDER: "Chủ TK:",
+    ACCOUNT_NUMBER: "Số tài khoản",
+    CREATE_DATE: "Ngày tạo",
+    ACCOUNT_HOLDER_FULL: "Chủ tài khoản",
+    NO_BANK_ACCOUNT: "Chưa có tài khoản ngân hàng nào",
+    CLICK_ADD_NEW_ACCOUNT: "Nhấn nút bên dưới để thêm tài khoản mới",
+  },
+  BUTTON: {
+    EDIT: "Chỉnh sửa",
+    DELETE: "Xóa",
+    ADD_NEW_BANK_ACCOUNT: "Thêm tài khoản ngân hàng",
+  },
+  MESSAGE_SUCCESS: {
+    DELETE_BANK_ACCOUNT: "Xóa tài khoản ngân hàng thành công",
+  },
+  MESSAGE_ERROR: {
+    BANK_ACCOUNT_NOT_FOUND: "Tài khoản ngân hàng không tồn tại",
+  },
+};
+export default TEXT_TRANSLATE_BANK_ACCOUNT;
