+++ app/(tabs)/(account)/bank-account/function-bank-account/FunctionBankAccount.translate.ts
@@ -1,0 +1,22 @@
+const TEXT_TRANSLATE_FUNCTION_BANK_ACCOUNT = {
+  TITLE: {
+    VALID_BANK_ACCOUNT_NOT_FOUND: "Không tìm thấy ngân hàng phù hợp",
+  },
+
+  BUTTON: {
+    UPDATE: "Cập nhật",
+    CONFIRM: "Xác nhận",
+    CHOOSE_BANK: "Chọn ngân hàng",
+  },
+
+  MESSAGE_SUCCESS: {
+    DELETE_BANK_ACCOUNT_SUCCESS: "Xóa tài khoản ngân hàng thành công",
+    UPDATE_BANK_ACCOUNT_SUCCESS: "Cập nhật tài khoản ngân hàng thành công",
+    ADD_BANK_ACCOUNT_SUCCESS: "Thêm tài khoản ngân hàng thành công",
+  },
+  MESSAGE_ERROR: {
+    BANK_ACCOUNT_NOT_FOUND: "Tài khoản ngân hàng không tồn tại",
+  },
+};
+
+export default TEXT_TRANSLATE_FUNCTION_BANK_ACCOUNT;
